+++ NTDLS.Katzebase.Engine/Parsers/StaticQueryParser.cs
@@ -68,6 +68,7 @@ namespace NTDLS.Katzebase.Engine.Parsers
                 QueryType.Begin => StaticParserBegin.Parse(queryBatch, tokenizer),
                 QueryType.Commit => StaticParserCommit.Parse(queryBatch, tokenizer),
                 QueryType.Rollback => StaticParserRollback.Parse(queryBatch, tokenizer),
+                QueryType.Create => StaticParserCreate.Parse(queryBatch, tokenizer),
 
                 _ => throw new KbParserException($"The query type is not implemented: [{token}]."),
             };
@@ -158,212 +159,6 @@ namespace NTDLS.Katzebase.Engine.Parsers
             }
             }
             */
-            #endregion
-
-            #region Create ---------------------------------------------------------------------------------------------
-
-            /*
-            else if (queryType == QueryType.Create)
-            {
-            if (tokenizer.TryIsNextToken(["schema", "index", "uniquekey", "procedure"]) == false)
-            {
-                throw new KbParserException("Invalid query. Found '" + tokenizer.PeekNext() + "', expected 'schema', 'index', 'uniquekey' or 'procedure'.");
-            }
-
-            token = tokenizer.GetNext();
-            if (Enum.TryParse(token, true, out SubQueryType subQueryType) == false)
-            {
-                throw new KbParserException("Invalid query. Found '" + token + "', expected: 'schema', 'index' or 'uniquekey'.");
-            }
-            result.SubQueryType = subQueryType;
-
-            token = tokenizer.GetNext();
-            if (token == string.Empty)
-            {
-                throw new KbParserException("Invalid query. Found '" + token + "', expected: object name.");
-            }
-
-            if (subQueryType == SubQueryType.Procedure)
-            {
-                result.AddAttribute(PreparedQuery.QueryAttribute.ObjectName, token);
-
-                var parameters = new List<PhysicalProcedureParameter>();
-
-                if (tokenizer.NextCharacter == '(') //Parse parameters
-                {
-                    tokenizer.SkipNextCharacter();
-
-                    while (true)
-                    {
-                        var paramName = tokenizer.GetNext();
-                        if (tokenizer.GetNext().Is("as") == false)
-                        {
-                            throw new KbParserException("Invalid query. Found '" + tokenizer.Breadcrumbs.Last() + "', expected: 'AS'.");
-                        }
-                        token = tokenizer.GetNext();
-
-                        if (Enum.TryParse(token, true, out KbProcedureParameterType paramType) == false || Enum.IsDefined(typeof(KbProcedureParameterType), paramType) == false)
-                        {
-                            string acceptableValues = string.Join("', '",
-                                Enum.GetValues<KbProcedureParameterType>().Where(o => o != KbProcedureParameterType.Undefined));
-
-                            throw new KbParserException($"Invalid query. Found '{token}', expected: '{acceptableValues}'.");
-                        }
-
-                        parameters.Add(new PhysicalProcedureParameter(paramName, paramType));
-
-                        if (tokenizer.NextCharacter != ',')
-                        {
-                            if (tokenizer.NextCharacter != ')')
-                            {
-                                throw new KbParserException("Invalid query. Found '" + tokenizer.NextCharacter + "', expected: ')'.");
-                            }
-                            tokenizer.SkipNextCharacter();
-                            break;
-                        }
-                        tokenizer.SkipNextCharacter();
-                    }
-                }
-
-                result.AddAttribute(PreparedQuery.QueryAttribute.Parameters, parameters);
-
-                if (tokenizer.GetNext().Is("on") == false)
-                {
-                    throw new KbParserException("Invalid query. Found '" + tokenizer.Breadcrumbs.Last() + "', expected: 'ON'.");
-                }
-
-                token = tokenizer.GetNext();
-                if (token == string.Empty)
-                {
-                    throw new KbParserException("Invalid query. Found '" + token + "', expected: schema name.");
-                }
-
-                result.AddAttribute(PreparedQuery.QueryAttribute.Schema, token);
-
-                if (tokenizer.GetNext().Is("as") == false)
-                {
-                    throw new KbParserException("Invalid query. Found '" + tokenizer.Breadcrumbs.Last() + "', expected: 'AS'.");
-                }
-
-                if (tokenizer.NextCharacter != '(')
-                {
-                    throw new KbParserException("Invalid query. Found '" + tokenizer.NextCharacter + "', expected: '('.");
-                }
-
-                if (tokenizer.Remainder().Last() != ')')
-                {
-                    throw new KbParserException("Invalid query. Found '" + tokenizer.NextCharacter + "', expected: ')'.");
-                }
-
-                tokenizer.SkipNextCharacter(); // Skip the '('.
-
-                var batches = new List<string>();
-
-                int previousPosition = tokenizer.Caret;
-
-                while (tokenizer.IsEnd() == false)
-                {
-                    if (tokenizer.NextCharacter == ')')
-                    {
-                        tokenizer.SkipNextCharacter();
-                    }
-                    else
-                    {
-                        _ = PrepareNextQuery(tokenizer);
-
-                        string queryText = tokenizer.Text.Substring(previousPosition, tokenizer.Caret - previousPosition).Trim();
-
-                        foreach (var literalString in tokenizer.StringLiterals)
-                        {
-                            queryText = queryText.Replace(literalString.Key, literalString.Value);
-                        }
-
-                        batches.Add(queryText);
-
-                        previousPosition = tokenizer.Caret;
-                        var nextToken = tokenizer.PeekNext();
-                    }
-                }
-
-
-
-                result.AddAttribute(PreparedQuery.QueryAttribute.Batches, batches);
-            }
-            else if (subQueryType == SubQueryType.Schema)
-            {
-                result.Schemas.Add(new QuerySchema(token));
-
-                if (tokenizer.TryIsNextToken("with"))
-                {
-                    var options = new ExpectedWithOptions
-                        {
-                            {"pagesize", typeof(uint) }
-                        };
-                    StaticWithOptionsParser.ParseWithOptions(ref tokenizer, options, ref result);
-                }
-            }
-            else if (subQueryType == SubQueryType.Index || subQueryType == SubQueryType.UniqueKey)
-            {
-                result.AddAttribute(PreparedQuery.QueryAttribute.IndexName, token);
-                result.AddAttribute(PreparedQuery.QueryAttribute.IsUnique, (subQueryType == SubQueryType.UniqueKey));
-
-
-                if (tokenizer.NextCharacter != '(')
-                {
-                    throw new KbParserException("Invalid query. Found '" + tokenizer.NextCharacter + "', expected: ','.");
-                }
-                tokenizer.SkipDelimiters('(');
-
-                while (true) //Get fields
-                {
-                    token = tokenizer.GetNext().ToLowerInvariant();
-                    if (token == string.Empty)
-                    {
-                        throw new KbParserException("Invalid query. Found '" + tokenizer.PeekNext() + "', expected: ',' or ')'.");
-                    }
-
-                    result.CreateFields.Add(token);
-
-                    if (tokenizer.NextCharacter == ',')
-                    {
-                        tokenizer.SkipDelimiters(',');
-                    }
-                    if (tokenizer.NextCharacter == ')')
-                    {
-                        tokenizer.SkipDelimiters(')');
-                        break;
-                    }
-                }
-
-                if (tokenizer.GetNext().Is("on") == false)
-                {
-                    throw new KbParserException("Invalid query. Found '" + token + "', expected: 'on'.");
-                }
-
-                token = tokenizer.GetNext();
-                if (!TokenHelpers.IsValidIdentifier(token, ':'))
-                {
-                    throw new KbParserException("Invalid query. Found '" + token + "', expected: schema name.");
-                }
-
-                result.Schemas.Add(new QuerySchema(token));
-
-                if (tokenizer.TryIsNextToken("with"))
-                {
-                    var options = new ExpectedWithOptions
-                        {
-                            {"partitions", typeof(uint) }
-                        };
-                    StaticWithOptionsParser.ParseWithOptions(ref tokenizer, options, ref result);
-                }
-            }
-            else
-            {
-                throw new KbNotImplementedException();
-            }
-            }
-            */
-
             #endregion
 
             #region Drop -----------------------------------------------------------------------------------------------
